refactor: 일급 컬렉션 리팩터링
