+++ src/main/java/model/Lotto.java
@@ -3,39 +3,21 @@ package model;
 import java.util.List;
 
 public class Lotto {
-    private final List<Integer> numbers;
+    private final LottoNumbers numbers;
     public static final int MIN_LOTTO_NUMBER = 1;
     public static final int MAX_LOTTO_NUMBER = 45;
     public static final int LOTTO_SIZE = 6;
 
     public Lotto(List<Integer> numbers) {
-        validate(numbers);
-        this.numbers = numbers;
+        this.numbers = new LottoNumbers(numbers);
     }
 
     public List<Integer> getNumbers() {
-        return numbers;
-    }
-
-    private void validate(List<Integer> lottoNumbers) {
-        validateNumbersSize(lottoNumbers);
-        validateNumbersRange(lottoNumbers);
+        return numbers.asList();
     }
 
     @Override
     public String toString() {
         return numbers.toString();
-    }
-
-    private void validateNumbersSize(List<Integer> lottoNumbers) {
-        if (lottoNumbers.size() != LOTTO_SIZE) {
-            throw new RuntimeException("로또 번호는 6개여야 합니다.");
-        }
-    }
-
-    private void validateNumbersRange(List<Integer> lottoNumbers){
-        if (lottoNumbers.stream().anyMatch(n -> n < MIN_LOTTO_NUMBER || n > MAX_LOTTO_NUMBER)) {
-            throw new RuntimeException("로또 번호는 1이상 45이하여야 합니다.");
-        }
     }
 }
